Server/requirements.txt
-i https://pypi.org/simple
asgiref==3.3.1; python_version >= '3.5'
attrs==20.3.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
base58==2.1.0; python_version >= '3.5'
bitarray==1.2.2
certifi==2020.12.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
cytoolz==0.11.0; implementation_name == 'cpython'
django-cors-headers==3.7.0
django==3.0.5
djangorestframework==3.12.2
djongo==1.3.4
eth-abi==2.1.1; python_version >= '3.6' and python_version < '4'
eth-account==0.5.4; python_version >= '3.6' and python_version < '4'
eth-hash[pycryptodome]==0.3.1; python_version >= '3.5' and python_version < '4'
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.1; python_version >= '3.6' and python_version < '4'
eth-typing==2.2.2; python_version >= '3.5' and python_version < '4'
eth-utils==1.10.0; python_version >= '3.5' and python_version < '4' and python_full_version != '3.5.2'
hexbytes==0.2.1; python_version >= '3.6' and python_version < '4'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
ipfs-api==0.2.3
ipfshttpclient==0.7.0a1; python_full_version >= '3.5.4' and python_full_version not in '3.6.0, 3.6.1, 3.7.0, 3.7.1'
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
netaddr==0.8.0
parsimonious==0.8.1
pillow==8.1.2
protobuf==3.15.6
pycryptodome==3.10.1
pymongo==3.11.3
pyrsistent==0.17.3; python_version >= '3.5'
python-magic==0.4.22
pytz==2021.1
requests==2.25.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
rlp==2.0.1
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.2.4; python_version >= '3.5'
toolz==0.11.1; python_version >= '3.5'
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
varint==1.0.2
web3==5.17.0
websockets==8.1; python_full_version >= '3.6.1'
